1.0.2 - Реализованы REST контроллеры(не работает JS)

+++ src/main/java/web/controller/AdminController.java
@@ -2,6 +2,9 @@ package web.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +16,7 @@ import web.service.UserService;
 import java.security.Principal;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Set;
 
 @RestController
 public class AdminController {
@@ -27,32 +31,48 @@ public class AdminController {
     }
 
     @GetMapping(value = "/admin")
-    public  List<User> showAllUsers() {
+    public ResponseEntity<List<User>> showAllUsers() {
         List<User> allUsers = userService.getAllUser();
-        return allUsers;
+        return new ResponseEntity<>(allUsers, HttpStatus.OK);
     }
 
     @PostMapping(value = "/admin")
-    public User createUser(@RequestBody User user) {
+    public ResponseEntity<User> createUser(@RequestBody User user) {
+        if (user == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         userService.addUser(user);
-        return user;
+        return new ResponseEntity<>(user, HttpStatus.CREATED);
     }
 
-    @GetMapping(value = "/admin/{id}")
-    public User getOneUser(@PathVariable("id") int id) {
+    @GetMapping(value = "/admin/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<User> getOneUser(@PathVariable("id") Integer id) {
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         User user = userService.getUserById(id);
-        return user;
+
+        if (user == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
     @PatchMapping("/admin")
-    public User updateUser(@RequestBody User user) {
+    public ResponseEntity<User> updateUser(@RequestBody User user) {
         userService.editUser(user);
-        return user;
+        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
     @DeleteMapping(value = "/admin/{id}")
-    public String deleteUser(@PathVariable int id) {
+    public ResponseEntity<?> deleteUser(@PathVariable Integer id) {
+
+        if (id == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         userService.deleteUser(userService.getUserById(id));
-        return "User with ID = " + id + " was deleted";
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
